github-mcp/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-mcp-adapters>=0.0.1
langgraph>=0.0.20
python-dotenv>=1.0.0

# API dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
requests>=2.31.0
